UHF-6849: Move IsSchool filter to custom module, replace duplicate solution
